refactor: renames textUtils into text
