+++ frontend/amundsen_application/static/js/utils/text.ts
